refactor(visibility): tidy warnings threading after simplify pass
- Move merge_visibility_warnings to tools/util_helpers.py (generic response
  helper) so visibility/resolver.py stays pure-resolver.
- Drop redundant set() copies on the two degraded-registry returns; denied is
  already a set and callers treat it read-only.
- Use the compositional merge in the domain-only search return for consistency
  with the other seams.
