+++ tests/src/unit/visibility/test_resolver.py
@@ -15,10 +15,7 @@ def _hidden(registry_result, config):
 
 def test_enabled_bad_payload_warns_and_fails_open(caplog):
     with caplog.at_level(logging.WARNING):
-        assert (
-            _hidden({"success": False}, VisibilityConfig(enabled=True))
-            == set()
-        )
+        assert _hidden({"success": False}, VisibilityConfig(enabled=True)) == set()
     assert any("unusable" in r.message for r in caplog.records)
 
 
@@ -26,19 +23,14 @@ def test_enabled_exception_payload_fails_open(caplog):
     # gather(return_exceptions=True) can hand an Exception object through as the
     # registry payload; it is non-dict, so the filter degrades to empty (open).
     with caplog.at_level(logging.WARNING):
-        assert (
-            _hidden(RuntimeError("boom"), VisibilityConfig(enabled=True))
-            == set()
-        )
+        assert _hidden(RuntimeError("boom"), VisibilityConfig(enabled=True)) == set()
     assert any("unusable" in r.message for r in caplog.records)
 
 
 def test_enabled_non_list_result_warns_and_fails_open(caplog):
     with caplog.at_level(logging.WARNING):
         assert (
-            _hidden(
-                {"success": True, "result": "nope"}, VisibilityConfig(enabled=True)
-            )
+            _hidden({"success": True, "result": "nope"}, VisibilityConfig(enabled=True))
             == set()
         )
     assert any("not a list" in r.message for r in caplog.records)
@@ -47,10 +39,7 @@ def test_enabled_non_list_result_warns_and_fails_open(caplog):
 def test_disabled_bad_payload_stays_silent(caplog):
     # Disabled is the default no-op; it must NOT warn on every call.
     with caplog.at_level(logging.WARNING):
-        assert (
-            _hidden({"success": False}, VisibilityConfig(enabled=False))
-            == set()
-        )
+        assert _hidden({"success": False}, VisibilityConfig(enabled=False)) == set()
     assert not caplog.records
 
 
@@ -157,10 +146,7 @@ def test_deny_honored_on_unusable_registry():
     assert _hidden({"success": False}, cfg) == {"sensor.ghost"}
     assert _hidden("nonsense", cfg) == {"sensor.ghost"}
     # dict-success payload but result is not a list
-    assert (
-        _hidden({"success": True, "result": "nope"}, cfg)
-        == {"sensor.ghost"}
-    )
+    assert _hidden({"success": True, "result": "nope"}, cfg) == {"sensor.ghost"}
 
 
 def test_degraded_registry_surfaces_warning():
